Implement SSSP with predecessors
The current SSSP implementation in cugraph computes only distances, not
predecessors. The predecessors are necessary for extracting the actual paths.
This commit provides a new SSSP implementation with can compute both distances
and predecessors.

The implementation is based on frontier expansion similar to BFS. The primary
differences compared to BFS are that:
 - Vertices can be visited multiple times during the course of the algorithm
 - Distances and weights are floats/doubles. Hence, in a given frontier
expansion, a vertex's distance may be updated mutiple times by multiple parents
In order to avoid duplicates in the frontier and to assign the right predecessors,
SSSP needs a two phase approach. The current implementation strikes a balance
between performance and the additional memory requirements. Further optimisations
that trade off space or time are possible.

Since this implementation shares code with BFS, common functions from BFS have
been factored out.

FREEBIE:
Fixed one memory overflow bug in the common BFS code