Build a Dendrogram class, adapt Louvain/Leiden/ECG to use it (#1359)
Preparing for MNMG Leiden and ECG identified an area for code cleanup.

The original cuGraph implementation of Louvain would flatten the hierarchical clustering as it was computed, filling (and returning) the final clustering.  This adds an awkward step in the middle of the Louvain computation.  Additionally, since Louvain (and Leiden and ECG which derive from it) is actually a hierarchical clustering algorithm it would be nice to generate the actual Dendrogram.

This PR implements a Dendrogram class, a function for flattening the Dendrogram, and modifies Louvain, Leiden and ECG to use the Dendrogram class.

It was suggested that the Dendrogram class could be moved to raft, decided to defer that until later, it's easy enough to move.

Authors:
  - Chuck Hastings (@ChuckHastings)

Approvers:
  - Alex Fender (@afender)
  - Brad Rees (@BradReesWork)
  - Andrei Schaffer (@aschaffer)
  - Rick Ratzel (@rlratzel)
  - AJ Schmidt (@ajschmidt8)

URL: https://github.com/rapidsai/cugraph/pull/1359