Create cuGraph developers guide (#1431)
We have long needed a developer's guide.  This PR creates one.  The expectation is that this document will be adapted over time.

I have copied the version from cuDF, edited things to make it cuGraph focused.  There are documentation holes here that will need to eventually be filled in.  However, it's probably useful to get something with holes in it that is approved so we can move forward with some documented standard.

Authors:
  - Chuck Hastings (@ChuckHastings)

Approvers:
  - Rick Ratzel (@rlratzel)
  - Brad Rees (@BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/1431