adding test graphs - part 2 (#1603)
This effort was originally targeted toward the WCC effort, but has been expanded a bit.  This supersedes #1545 which I will close.

The goal here is to create a means for constructing test graphs in an easier fashion.  Testing the capabilities of different graph algorithms might require a variety of graphs.  The objective of this PR is to better organize the graph generation components and to introduce some components to help in composing graphs out of multiple components.

This PR introduces the following capabilities:
* Create an ER graph
* Create a collection of Complete Graphs
* Create a collection of 2D mesh graphs
* Create a collection of 3D mesh graphs
* Create a random path graph (connect all vertices with a single randomly ordered path)
* Translate vertex ids of a graph
* Combine multiple edge lists into a single graph

Closes #1543

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Andrei Schaffer (https://github.com/aschaffer)
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/1603