revert cuco to latest dev branch, issues should be fixed (#1721)
We had an issue in CI that uncovered the fact that raft is referencing the `dev` branch of coco while `cugraph` has it pinned to a previous commit.  CI is using the newer `dev` branch version, which doesn't match what happens when we do local builds.

The original motivation for pinning was a temporary fix because of a bug introduced in `coco` which has since been resolved.  This PR reverts us back to using the `dev` branch.

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/1721