Updated to handle CUDA version check on machines with no GPUs (#1777)
Updated to handle CUDA version check on machines with no GPUs. When no GPUs are present, cugraph can import and ktruss is available, which should match the original behavior prior to when the checks were added for disabling ktruss on CUDA 11.4

Tested by setting CUDA_VISIBLE_DEVICES to " ", verifying the expected exception was raised, made the change, then verified cugraph could be imported.