Disable WCC test until we get get on an A100 to debug on (#1870)
CI testing on A100 is detecting a segmentation fault in the WCC implementation.  We currently can't get developer access to an A100 in order to debug this.  This PR will disable the WCC test.  Once we can resolve the A100 issue we can enable this test again.

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)
  - Rick Ratzel (https://github.com/rlratzel)

Approvers:
  - Brad Rees (https://github.com/BradReesWork)
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/cugraph/pull/1870