initial creation of libcugraph_etl.so (#1885)
The new renumbering implementation will require C++ integration directly with cudf.  In order to facilitate that, but also support our customers that won't need cudf, this PR will create a separate library (`libcugraph_etl.so`) which will ultimately link with `libcudf.so` and contain the ETL portions of cugraph that require cudf features.

This way our other libcugraph customers that don't need to reference the new library will not need to install all of the cudf dependencies.

To seed this, the PR also includes a proposed API for the new renumbering capability.

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)
  - Rick Ratzel (https://github.com/rlratzel)

Approvers:
  - Seunghwa Kang (https://github.com/seunghwak)
  - https://github.com/chirayuG-nvidia
  - Rick Ratzel (https://github.com/rlratzel)
  - Jordan Jacobelli (https://github.com/Ethyling)

URL: https://github.com/rapidsai/cugraph/pull/1885