Revert SG Katz API's signature to previous <22.04 version (#2242)
This PR is a slight reversion of a minor change from #2201, which changed the signature for the refactored SG Katz method in cugraph. That PR was breaking, but when combined with this one, will be non-breaking for the next release. This closes #1759 (already closed).

Authors:
  - https://github.com/betochimas

Approvers:
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/2242