Renumber PG to be contiguous per type (#2697)
Closes #2627 and CC @VibhuJawa 

Currently only SG as we hammer out the API and behavior.

This returns a dataframe with start and stop for each type.  Should stop be inclusive or exclusive?

How should we handle vertex ids that only exist in edge data?  Should we raise (for now) if this condition exists?  I think we can handle this without _too_ much difficulty, but it will take more work.

Since we number edge data, I think edge IDs will often be added in a way that is already contiguous per type.  We could keep track of this to avoid unnecessary computation.

Also, I want to confirm that we _cannot_ have multiple rows for a single vertex ID, right?  I think we settled on this.  Multiple rows with the same ID would cause a problem with the current implementation--it currently gives each row a unique ID.

Authors:
  - Erik Welch (https://github.com/eriknw)

Approvers:
  - Vibhu Jawa (https://github.com/VibhuJawa)
  - Alex Barghi (https://github.com/alexbarghi-nv)
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/2697