Fix docs builds
In #2684, the CMake build system was switched from `make` to `ninja`, which broke these doc builds.

This PR updates the scripts for the nightly docs builds to use a generic `cmake` build option, rather than a build-system specific invocation to fix the issue and prevent a similar issue from occurring in the future.

**Note**: this PR targets `branch-22.10` since the the `22.10` docs aren't currently up-to-date because of this issue.