Add vector properties (#2882)
~WIP.~ This ~probably~ closes #2721.

This uses cudf List dtypes to store vectors. When converting a vector column to arrow, the data appears to be on host, so it's unclear how many copies and moves of the data we're doing, but I don't think we have many easy alternatives besides relying on what cudf gives us.  In pandas, vector properties are object dtype and stored as numpy arrays.

I think it makes sense for a vector property to be required to be the same length (i.e., if it's added multiple times).

We may want to add a method to convert a vector property to a numpy or cupy array.
- How do we handle null values? Raise? Set to 0? Ignore/skip?
- Should users be able to specify numpy or cupy array (host or device)?

When getting data, should we allow vector properties to be expanded?

Can we create a graph with vector property data?

Should we add a keyword argument to `add_vertex_data` to say "use all (or the rest) of columns as a vector property of this name"?

Should we allow vector properties to come in already as cupy List dtype?

Authors:
  - Erik Welch (https://github.com/eriknw)

Approvers:
  - Brad Rees (https://github.com/BradReesWork)
  - Alex Barghi (https://github.com/alexbarghi-nv)
  - Vibhu Jawa (https://github.com/VibhuJawa)

URL: https://github.com/rapidsai/cugraph/pull/2882