CuGraph-Service Asyncio Fix (#2932)
Resolves #2875 

This resolves a bug where the CGS client tried to start a new event loop without checking whether it was in an existing event loop.  Running a new event loop on the same thread as an existing loop is not allowed, and this caused CGS to fail on Jupyter when using the CGS `result_device` option.

This PR resolves the bug by creating a new thread when necessary so that CGS can run its own event loop when one is already running in the current thread.

Authors:
  - Alex Barghi (https://github.com/alexbarghi-nv)

Approvers:
  - Brad Rees (https://github.com/BradReesWork)
  - Vibhu Jawa (https://github.com/VibhuJawa)

URL: https://github.com/rapidsai/cugraph/pull/2932