Support cuGraph-Service in cuGraph-PyG (#2946)
Resolves https://github.com/rapidsai/graph_dl/issues/33

This PR resolves the most critical issue for `cuGraph-PyG`, by adding support for `cuGraph-Service`.  `cuGraph-PyG` supports `cuGraph-Service` through the `RemoteGraph` API.  As a result of this API standardization, there are few major changes to `cugraph_store`, `dispatch`, and `cugraph_sampler`.  Instead, this PR primarily addresses some edge cases that only appear with remote graphs.

This PR also contains a bug fix for cuGraph-Service, where the `GraphVertexEdgeID` thrift union was not being created correctly, and an associated test update.  This bug was found during testing of cuGraph-Service+PyG.

This PR also contains a bug fix for cuGraph-Service, where cupy was mistakenly included as a required dependency.  It has now been made an optional (soft) dependency using `import_optional`.

This PR also adds a new notebook that demonstrates an end-to-end workflow using the PyG extensions and cuGraph-Service.

I also apologize for the large number of commits; this is due to the other PRs that this PR depends on not being merged.

Authors:
  - Alex Barghi (https://github.com/alexbarghi-nv)
  - Erik Welch (https://github.com/eriknw)

Approvers:
  - Joseph Nke (https://github.com/jnke2016)
  - Rick Ratzel (https://github.com/rlratzel)
  - Vibhu Jawa (https://github.com/VibhuJawa)

URL: https://github.com/rapidsai/cugraph/pull/2946