Updates to allow python benchmarks to run on additional datasets by default (#3506)
closes #3321
closes #3331

This PR includes updates to allow benchmarks to run on additional datasets that don't have weight columns, and to auto convert to undirected for algos that require that.

These datasets are not normally run in CI (when CI runs them as tests), since CI uses the `-m tiny` marker to get coverage using a few small datasets in order to keep CI runs quick.

This also removes the `--no-rmm-reinit` pytest option and replaces it with `--allow-rmm-reinit` in order to create a default set of RMM options to run with, instead of all combinations by default.  A fixed set of options is (according to my understanding) the typical use case, and changing them multiple times within the same process is atypical and not supported well.  Testers now have to pass `--allow-rmm-reinit` to have pytest run with multiple RMM config options.

Authors:
  - Rick Ratzel (https://github.com/rlratzel)

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)
  - Alex Barghi (https://github.com/alexbarghi-nv)

URL: https://github.com/rapidsai/cugraph/pull/3506