Add support in C API for handling unweighted graphs in algorithms that expect weights (#3513)
Closes #3499 

Some of our C++ algorithms require a graph to be weighted.  The python layer will add "fake" edge weights of 1 to every edge in order to call these algorithms.  This adds a persisted constant value for every edge that will continue to exist for the life span of the python graph object, even though it is only necessary for a handful of algorithms.

This PR addresses these concerns.

Many of the C++ algorithms already actually support the graph not having weights.  A few should support it and had bugs that were corrected (induced subgraph, k-core, egonet).  A few can easily support it but did not (legacy spectral algorithms).

The Louvain and Leiden algorithms need to have a weighted graph because they will coarsen the weights as the graph is coarsened, and the weights will eventually not be one while the algorithm is running.  These require a slightly different approach.

This PR also added C API tests for unweighted graphs for these algorithms.

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)
  - Brad Rees (https://github.com/BradReesWork)

Approvers:
  - Seunghwa Kang (https://github.com/seunghwak)

URL: https://github.com/rapidsai/cugraph/pull/3513