[BUG] Fix Bulk Sampling Test Issue (#3701)
The bulk sampling tests are currently using a temporary directory which is not accessible from all processes/workers.  This is causing the MG tests to fail.

This PR updates the bulk sampler tests to use an environment variable for the scratch directory.  The test run script will also need to be updated to make sure that environment variable is properly set.

Authors:
  - Alex Barghi (https://github.com/alexbarghi-nv)

Approvers:
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/3701