Clean-up old testing conventions in `test_ecg.py` (#3779)
Closes #3235 

This PR cleans up the `test_ecg.py` unit test. It removes the old way of comparing networkx results to cugraph, which used `PurePath` to compare the location of the stored dataset when getting networkx results to compare with cugraph.

Since the cugraph tests use the `Datasets` API, the golden results can be fetched based on the name of the dataset being used.

Authors:
  - ralph (https://github.com/nv-rliu)

Approvers:
  - Alex Barghi (https://github.com/alexbarghi-nv)

URL: https://github.com/rapidsai/cugraph/pull/3779