Allow models to use a lightweight sparse structure (#3782)
This PR introduces `SparseGraph` class to allow `SAGEConv` to use a more lightweight graph structure. The goal is to provide an option to bypass `to_block` which is the bottleneck in the sampling workflow.

I will submit another PR to extend the pattern to other models.

Authors:
  - Tingyu Wang (https://github.com/tingyu66)
  - Alex Barghi (https://github.com/alexbarghi-nv)

Approvers:
  - Vibhu Jawa (https://github.com/VibhuJawa)
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/3782