Fix subtle memory leak in nbr_intersection primitive (#3858)
Closes https://github.com/rapidsai/graph_dl/issues/259

A customer found a subtle memory leak in Jaccard similarity.  Tracked it down to this subtle error.

`major_nbr_indices` is an `std::optional` that is initialized to `std::nullopt`.  Overwriting the dereferenced entry replaces the value but does not mark the optional as containing a value.  So the resulting value is never destroyed.

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Seunghwa Kang (https://github.com/seunghwak)

URL: https://github.com/rapidsai/cugraph/pull/3858