[BUG] Always Persist Dask DataFrames in cuGraph-DGL Graph Storage (#4296)
Always persists the dask dataframe in the cuGraph-DGL graph storage object.  This resolves a bug where the dataframe was unpersisted, causing `'TypeError("Could not construct DataFrame from <class \'tuple\'>")'` to be raised when trying to access it after saving it as a dataset.

Authors:
  - Alex Barghi (https://github.com/alexbarghi-nv)

Approvers:
  - Vibhu Jawa (https://github.com/VibhuJawa)
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/4296