Expose graph_view method to count multi edges through C API and PLC (#4426)
Python property graph class does some expensive dask calls to determine if a graph is a multi-graph.

This PR exposes to the C API and the PLC layer the graph_view method that counts the number of multi-edges in the graph.  Specifically, the function returns the number of extra edges exist.  So if the edge `(u,v)` exists `k` times, that edge contributes `k-1` to the return value.

If the return value of this function is greater than 0 then the graph is a multi-graph.  If the return value is 0 then while the graph might be labeled a multi-graph, there are no multi-edges so it could be treated as a graph.

Closes #2417 

Note that the property graph python code would still need to be modified to use this function.

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Seunghwa Kang (https://github.com/seunghwak)
  - Joseph Nke (https://github.com/jnke2016)
  - Alex Barghi (https://github.com/alexbarghi-nv)

URL: https://github.com/rapidsai/cugraph/pull/4426