[IMP] Limit the Test Data Size when Running CI in `gcn_dist_sg.py` (#4461)
Running `gcn_dist_sg.py` can take 15-25 minutes in CI because of the size of the test dataset.  This PR cuts the size of that test dataset for CI runs, which results in nearly identical accuracy anyways and a runtime of less than 5 minutes.

Authors:
  - Alex Barghi (https://github.com/alexbarghi-nv)
  - Naim (https://github.com/naimnv)

Approvers:
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/4461