nx-cugraph: add `G.__networkx_cache__` to enable graph conversion caching (#4567)
Adding this now, because this was an oversight.

By having a `G.__networkx_cache__` MutableMapping, we give NetworkX the ability to cache graph conversions. This will be most useful in dev and future versions of NetworkX. As such, it would be nice (but not strictly essential) to get this in 24.08.

Authors:
  - Erik Welch (https://github.com/eriknw)

Approvers:
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/4567