Set CUDA_STATIC_MATH_LIBRARIES in Python builds (#4612)
In https://github.com/rapidsai/cugraph/pull/4526, we introduced the CUDA_STATIC_MATH_LIBRARIES option and set it to OFF by default. We did not set it to ON in the Python builds, and as a result ended up bundling dynamic CUDA libraries. Set the option to ON in the Python build.

Authors:
  - Kyle Edwards (https://github.com/KyleFromNVIDIA)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cugraph/pull/4612