Adds new dataset for benchmarking in the 100k node 1M edge range, adds additional k-values for BC benchmarks (#4726)
Adds new dataset for benchmarking in the 100k node 1M edge range.
This also updates the benchmark fixture to download the dataset instead of requiring a separate script.
This also includes changes from #4721 to add additional k-values for BC benchmarks to run. 

The addition was done with an inline yaml file and a followup PR will be done to properly add this dataset to the cugraph.datasets API.  The inline yaml was done in order to allow benchmarks to run on a system with an existing cugraph installation by simply copying this bench_algos.py, rather than require an updated cugraph install with the new dataset metadata.

Authors:
  - Rick Ratzel (https://github.com/rlratzel)
  - Ralph Liu (https://github.com/nv-rliu)

Approvers:
  - Ralph Liu (https://github.com/nv-rliu)
  - Don Acosta (https://github.com/acostadon)

URL: https://github.com/rapidsai/cugraph/pull/4726