Don't compile int32_t/int64_t vertex_t/edge_t combinations (#4720)
To reduce compile time and library size... we're dropping pre-compiled support for vertex_t as int32_t and edge_t as int64_t.  This is a small edge case in memory utilization.

The data structures and primitives would still support such a configuration, but we're not compiling it into libcugraph.so with this PR. 

Seems to improve C++ compile time by 25% on my dgx18 and a 30% reduction in the size of libcugraph.so.

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Seunghwa Kang (https://github.com/seunghwak)

URL: https://github.com/rapidsai/cugraph/pull/4720