nx-cugraph: faster `shortest_path`
For larger graphs, nearly all the time was spent creating the dict of lists of paths.
I couldn't find a better way to create these, nor could I find an approach to compute
more in PLC or cupy. So, the solution in this PR is to avoid computing until needed!

This now returns a `Mapping` instead of a `dict`. Will anybody care or notice that
the return type isn't strictly a dict?

This is currently only for unweighted bfs. We should also do weighted sssp paths.

Also, this currently recurses. If we like the approach in this PR, we should update
computing the paths on demand to not recurse.