Fix CCCL 2.7.0-rc2 compile issue by removing reference from values. (#4799)
We are planning to migrate to CCCL 2.7.0 in RAPIDS 25.02. I found a compilation error in `libcugraph_etl` that is fixed by removing a reference `&` from a device comparator, and passing the parameters by value instead.

This is a pretty small fix but I expect we will want to build 24.12 with newer CCCL versions for DLFW releases. To avoid needing a backport for DLFW, I am targeting 24.12 with this change. I can retarget to 25.02 if desired.

This is extracted from #4743, but that PR shouldn't be merged (it includes testing changes that are not intended for merge).

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Chuck Hastings (https://github.com/ChuckHastings)

URL: https://github.com/rapidsai/cugraph/pull/4799