fix(rattler): add librmm to host to fix overlinking errors (#5065)
## Description

We need `librmm` for _building_ `libcudf` and `librmm` is available in
the cache build stage. However the overlinking checks happen on a
per-output basis and now that `librmm` isn't header-only, `rattler` is
detecting a missing link.
This adds the `librmm` entry to the `host` section of the appropriate
outputs.

xref rapidsai/build-planning#175