Fix Resultset Tests for Compat with `xdist` (#5125)
## Suggested Changes

This PR corrects an outdated unit test that was using custom TemporaryDirectories that would introduce transient failures when running CI with `pytest-xdist`. Now the test should use pytest's built-in `tmp_path` fixtures to ensure workers do not override variables across other workers. 

## Note

A `FIXME` was left in the test to signal that `testing/resultset.py` should be refactored. It is currently using multiple copies of the `DefaultDownloadDir` object in a way that easily introduces bugs. This will be addressed in a follow-up PR shortly

Authors:
  - Ralph Liu (https://github.com/nv-rliu)

Approvers:
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/5125