Update MST test (#5143)
A change in the Raft API was merged to add support for graph with identical vertices. However, because of the new random weights being added to each edge weight to break-tie, it leads to a different but valid MST. The CAPI test hence has a different hardcoded MST result. This PR updates the test accordingly.

Authors:
  - Joseph Nke (https://github.com/jnke2016)

Approvers:
  - Chuck Hastings (https://github.com/ChuckHastings)

URL: https://github.com/rapidsai/cugraph/pull/5143