fix pytorch requirement for CUDA 13 builds (#5244)
Follow-up to #5236

There, I intended to introduce a dependency on `torch` nightly wheels for CUDA 13, but made a typo in the `dependencies.yaml` filter. This fixes that, so that wheels for the correct `torch` variant + version will be installed in wheel-testing CI jobs here.

Authors:
  - James Lamb (https://github.com/jameslamb)

Approvers:
  - Gil Forsyth (https://github.com/gforsyth)

URL: https://github.com/rapidsai/cugraph/pull/5244