Re-enable spectral clustering (#5151)
This PR re-enables the legacy spectral clustering test. After a thorough discussion with the raft team, we concluded that because of the small graph size being tested, the algorithm is subject to more variation because of its random nature. Hence this test only ensures that the algorithm effectively runs while maintaining an edge_cut value lower than a certain threshold. 

This PR also

1. Update the C, PLC and Python API to support random state
2. Throw an exception if any of the legacy algorithms is called with a non-symmetric graph
3. Update the tests accordingly

Note: The raft API takes a seed instead of a random state as input hence in order to have the CAPI and PLC API align with other algorithms leveraging random state, we made a subtle change in the C++ API which generate a seed from the random state passed by the user. And that seed will be passed to the `raft` API. Once a C++ implementation leveraging our graph primitives is available, we will only need to update the C++ API and key the upper stream API unchaged. 

closes #4173

Authors:
  - Joseph Nke (https://github.com/jnke2016)

Approvers:
  - Chuck Hastings (https://github.com/ChuckHastings)
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/5151