[BUG] Fix Sampling Output Parameters for Temporal Sampling API (#5267)
Two of the heterogeneous+temporal sampling API calls in `pylibcugraph` use the wrong output parameters.  This PR restores the correct parameters that are required for processing heterogeneous graphs.  It also fixes a missing comma in a function signature that caused an error when trying to pass the `num_edge_types` parameter.

Authors:
  - Alex Barghi (https://github.com/alexbarghi-nv)
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Brad Rees (https://github.com/BradReesWork)
  - Chuck Hastings (https://github.com/ChuckHastings)

URL: https://github.com/rapidsai/cugraph/pull/5267