`force_atlas2`: allow vertex mass to be given by the user (#5293)
This is a "quick" follow-up to https://github.com/rapidsai/cugraph/pull/5260 (and https://github.com/rapidsai/cugraph/pull/5213) and https://github.com/rapidsai/nx-cugraph/pull/191 to add vertex mass to the API. This will allow `nx-cugraph` to better match the networkx API.

This required changing the dtype of the mass arrays from integral to float32.

As time permits, I plan to add a test to show that increasing mass of a vertex increases its distance to other vertices. The plan is to also update `nx-cugraph` to use this parameter.

Authors:
  - Erik Welch (https://github.com/eriknw)

Approvers:
  - Chuck Hastings (https://github.com/ChuckHastings)
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/5293