Time filtering hop 0 (#5328)
Added support for time filtering in hop 0 of temporal sampling.  If starting times are provided in the API for temporal sampling then we do filtering in that hop, if they are not provided then time filtering only occurs later (which is the previous behavior).

Refactored the C API testing for sampling as things were getting to complicated.

Closes #5263 

Change is breaking since we are changing the public API for temporal sampling and deleting some previously deprecated functions.

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Alex Barghi (https://github.com/alexbarghi-nv)
  - Joseph Nke (https://github.com/jnke2016)
  - Seunghwa Kang (https://github.com/seunghwak)

URL: https://github.com/rapidsai/cugraph/pull/5328